Add mocha metrics implementation
Note: Based on the tests we ran, the metrics returned by the HuggingFace
datasets and the ones returned by COCO are not exactly the same. This
may be due to some parameter on the metrics or due to differences in
preprocessing. Since we're normalizing and lowering case, it might be
the case that our estimates are over confident