Add compute_correlations_mocha.ipynb
This file contains the computation of Pearson and Spearman correlations
for the different metrics. The goal is to serve as a preliminary results
towards whether the human judgement scores are predictable or not from
the other metrics.

It also considers a simple case where a linear regression model is
fitted on the computed metrics.